Use current session to create a resume state instead of using the player
