+++ Sources/Castor/Cast/Cast.swift
@@ -5,6 +5,7 @@
 //
 
 import Combine
+import CoreMedia
 import Foundation
 import GoogleCast
 import SwiftUI
@@ -193,7 +194,7 @@ extension Cast: @preconcurrency GCKSessionManagerListener {
 
     // swiftlint:disable:next missing_docs
     public func sessionManager(_ sessionManager: GCKSessionManager, willEnd session: GCKCastSession) {
-        if let delegate, let resumeState = player?.resumeState(with: delegate) {
+        if let delegate, let resumeState = session.remoteMediaClient?.resumeState(with: delegate) {
             delegate.castEndSession(with: resumeState)
         }
     }
