+++ Sources/Castor/Player/CastPlayer.swift
@@ -79,10 +79,4 @@ public final class CastPlayer: NSObject, ObservableObject {
         __targetSeekTime.bind(to: remoteMediaClient)
         __items.bind(to: remoteMediaClient)
     }
-
-    func resumeState(with delegate: CastDelegate) -> CastResumeState? {
-        guard let mediaStatus = _mediaStatus, let currentIndex = mediaStatus.currentIndex() else { return nil }
-        let assets = mediaStatus.items().compactMap { delegate.castAsset(from: .init(rawMediaInformation: $0.mediaInformation)) }
-        return CastResumeState(assets: assets, index: currentIndex, time: time() - seekableTimeRange().start)
-    }
 }
